feat(fx-engine): extract pixi scenes and extend scene ids
