+++ packages/fx-engine/src/scenes/builtin/rank-promoted.scene.ts
@@ -1,0 +1,19 @@
+import type { FxEvent, FxRarity } from '../../types';
+import type { FxSceneDefinition } from '../contracts';
+
+export const rankPromotedScene: FxSceneDefinition<'rank.promoted'> = {
+  id: 'rank.promoted',
+  priority: 20,
+  supports: (event: FxEvent) => event.type === 'rank.promoted',
+  create: (event: FxEvent, rarity: FxRarity) => ({
+    sceneId: 'rank.promoted',
+    rarity,
+    payload: {
+      ...event,
+      type: 'rank.promoted',
+      subtitle: event.subtitle ?? 'PROMOTION UNLOCKED',
+      icon: event.icon ?? 'shield',
+      rarity,
+    },
+  }),
+};
