+++ packages/fx-engine/src/types.ts
@@ -2,13 +2,37 @@ export type FxMode = 'off' | 'lite' | 'epic';
 
 export type FxRarity = 'common' | 'rare' | 'epic' | 'legendary';
 
-export interface AchievementFxEvent {
+export type FxEventType =
+  | 'achievement.unlocked'
+  | 'rank.promoted'
+  | 'weapon.unlocked';
+
+interface FxEventBase {
   id?: string;
   title: string;
   subtitle?: string;
   icon?: string;
   rarity?: FxRarity;
+  metadata?: Record<string, unknown>;
 }
+
+export interface AchievementFxEvent extends FxEventBase {
+  type?: FxEventType;
+}
+
+export interface RankPromotedFxEvent extends FxEventBase {
+  type: 'rank.promoted';
+  rankTitle?: string;
+  rankLevel?: number;
+}
+
+export interface WeaponUnlockedFxEvent extends FxEventBase {
+  type: 'weapon.unlocked';
+  weaponName?: string;
+  weaponClass?: string;
+}
+
+export type FxEvent = AchievementFxEvent | RankPromotedFxEvent | WeaponUnlockedFxEvent;
 
 export interface FxRenderOptions {
   mode: FxMode;
@@ -17,7 +41,7 @@ export interface FxRenderOptions {
 }
 
 export interface FxRenderer {
-  playAchievement(event: AchievementFxEvent, options: FxRenderOptions): Promise<void>;
+  playAchievement(event: FxEvent, options: FxRenderOptions): Promise<void>;
   dispose(): void;
 }
 
@@ -28,7 +52,7 @@ export interface FxAudioService {
 }
 
 export interface FxController {
-  showAchievement(event: AchievementFxEvent): void;
+  showAchievement(event: FxEvent): void;
   setMode(mode: FxMode): void;
   setSafeMoment(value: boolean): void;
   unlockAudio(): void;
