+++ packages/fx-engine/tests/scene-runtime.test.ts
@@ -37,4 +37,40 @@ describe('renderScene', () => {
       options,
     );
   });
+
+  it('falls back to renderer for custom scene ids', async () => {
+    const playAchievementMock = jest.fn<FxRenderer['playAchievement']>(async () => undefined);
+    const renderer: FxRenderer = {
+      playAchievement: playAchievementMock,
+      dispose: jest.fn(),
+    };
+
+    await renderScene(
+      renderer,
+      {
+        sceneId: 'custom.special',
+        rarity: 'legendary',
+        payload: {
+          title: 'Custom Scene',
+          rarity: 'legendary',
+        },
+      },
+      {
+        mode: 'epic',
+        reducedMotion: false,
+      },
+    );
+
+    expect(playAchievementMock).toHaveBeenCalledTimes(1);
+    expect(playAchievementMock).toHaveBeenCalledWith(
+      {
+        title: 'Custom Scene',
+        rarity: 'legendary',
+      },
+      {
+        mode: 'epic',
+        reducedMotion: false,
+      },
+    );
+  });
 });
